% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_crtNwk}
\alias{fun_crtNwk}
\title{Create networks - outdated method}
\usage{
fun_crtNwk(nNodes, shape = c("lin", "int", "den"))
}
\arguments{
\item{nNodes}{Intiger numbr of nodes.}

\item{shape}{Network shape, one of "lin", "int" or "den".}
}
\value{
A list containing adjacency matrices and an igraph object.
}
\description{
Creates a DEN based on probability triplets defined by `shape` parameter
}
\examples{
\dontrun{
fun_crtNwk(10, "den")
}
}
