% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_crtNwk2}
\alias{fun_crtNwk2}
\title{Create networks - update method}
\usage{
fun_crtNwk2(nNodes, prb)
}
\arguments{
\item{nNodes}{Integer numbr of nodes.}

\item{prb}{Branching probability}
}
\value{
An adjacency matrix
}
\description{
Creates a DEN,
}
\examples{
\dontrun{
fun_crtNwk2(10, 0.5)
}
}
