% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_downstream_nodes}
\alias{fun_downstream_nodes}
\title{find all downstream nodes of a vertex
taken and edited from https://github.com/robertness/lucy/blob/master/R/lucy.R}
\usage{
fun_downstream_nodes(g, w)
}
\arguments{
\item{g}{a dendritic network as an igraph object}

\item{w}{target node}
}
\value{
a vector of downstream node id's
}
\description{
find all downstream nodes of a vertex
taken and edited from https://github.com/robertness/lucy/blob/master/R/lucy.R
}
\examples{
\dontrun{
fun_downstream_nodes(g, 5)
}
}
