% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_get_cut_off}
\alias{fun_get_cut_off}
\title{used internally by fun_shreve_mag}
\usage{
fun_get_cut_off(dwnAdj, jAdded, g)
}
\arguments{
\item{dwnAdj}{directed adjacency mtx}

\item{jAdded}{Vector of just added nodes}

\item{g}{graph object}
}
\value{
integer - cut-off distance
}
\description{
used internally by fun_shreve_mag
}
