% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_headW}
\alias{fun_headW}
\title{used internally by fun_shreve_mag}
\usage{
fun_headW(g)
}
\arguments{
\item{g}{graph object}
}
\value{
vector of "true" headwater nodes
}
\description{
used internally by fun_shreve_mag
}
