% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_headwater_nodes}
\alias{fun_headwater_nodes}
\title{get all headwater nodes}
\usage{
fun_headwater_nodes(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
a vector of headwater node id's
}
\description{
get all headwater nodes
}
\examples{
\dontrun{
fun_headwater_nodes(g)
}
}
