% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_pltNwk}
\alias{fun_pltNwk}
\title{Quick and dirty plotting of dendritic networks, wrapper for
igraph::layout_as_tree and plot.igraph.}
\usage{
fun_pltNwk(g, direct = c("in", "out"), ...)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
plot
}
\description{
Quick and dirty plotting of dendritic networks, wrapper for
igraph::layout_as_tree and plot.igraph.
}
\examples{
\dontrun{
fun_pltNwk(g)
}
}
