% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_reverse_graph}
\alias{fun_reverse_graph}
\title{reverse direction of a directed graph}
\usage{
fun_reverse_graph(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
an igraph object
}
\description{
reverse direction of a directed graph
}
\examples{
\dontrun{
fun_reverse_graph(g)
}
}
