% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_shreve_mag}
\alias{fun_shreve_mag}
\title{Shreves stream magnitude}
\usage{
fun_shreve_mag(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
an igraph object with stream Shreve's order added as an attribute
}
\description{
Shreves stream magnitude
}
