% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_strahler_order}
\alias{fun_strahler_order}
\title{Get strahler order}
\usage{
fun_strahler_order(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
an igraph object with stream strahler order added as an attribute
}
\description{
Get strahler order
}
\examples{
\dontrun{
fun_strahler_order(g)
}
}
