% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_upstream_nodes}
\alias{fun_upstream_nodes}
\title{find all upstream nodes of a vertex
taken and edited from https://github.com/robertness/lucy/blob/master/R/lucy.R}
\usage{
fun_upstream_nodes(g, w)
}
\arguments{
\item{g}{a dendritic network as an igraph object}

\item{w}{target node}
}
\value{
a vector of upstream node id's
}
\description{
find all upstream nodes of a vertex
taken and edited from https://github.com/robertness/lucy/blob/master/R/lucy.R
}
\examples{
\dontrun{
fun_upstream_nodes(g, 5)
}
}
