% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{generate.noise}
\alias{generate.noise}
\title{Generate noise signal}
\usage{
generate.noise(alpha = 0, beta = 1, C = 1, L = 0, H = 1, num_gens)
}
\arguments{
\item{alpha}{autocorrelation coefficient}

\item{beta}{xx}

\item{C}{xx}

\item{L}{Lower limit for rescaling, default = 0}

\item{H}{Upper limit for rescaling, default = 1}

\item{num_gens}{sequence length}
}
\value{
A vector
}
\description{
Generate noise signal
}
\examples{
\dontrun{
generate.noise(alpha = 0, beta = 1, C = 1, L = 0, H = 1, num_gens = 100)
}
}
