% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{rescale.to.range2}
\alias{rescale.to.range2}
\title{Rescale vector to desired range.}
\usage{
rescale.to.range2(x, L, H)
}
\arguments{
\item{x}{vector}

\item{L}{Lower limit}

\item{H}{Upper limit}
}
\value{
A vector
}
\description{
Rescale vector to desired range.
}
\examples{
\dontrun{
rescale.to.range2(seq(1, 20), 100, 200)
}
}
