% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{sample.vec}
\alias{sample.vec}
\title{used internally by fun_shreve_mag}
\usage{
sample.vec(x, ...)
}
\arguments{
\item{x}{vector of 1+ length}
}
\value{
vector
}
\description{
used internally by fun_shreve_mag
}
