% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{fun_pltNwk}
\alias{fun_pltNwk}
\title{quick and dirty plot of network}
\usage{
fun_pltNwk(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
plot
}
\description{
quick and dirty plot of network
}
\examples{
\dontrun{
fun_pltNwk(g)
}
}
